fintech-payments-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
slowapi==0.1.9
